fix: Diving by zero when calculating nps for bench
BENCH: 109712035

+++ src/search/bench.cpp
@@ -29,5 +29,7 @@ void search::bench::run(searcher& searcher, const u32 depth) {
     const auto seconds = elapsed / 1000;
 
     std::cout << std::format("\ninfo string {} seconds", seconds) << std::endl;
-    std::cout << std::format("{} nodes {} nps", total_nodes, total_nodes / seconds) << std::endl;
+    std::cout << std::format("{} nodes {} nps", total_nodes,
+                             static_cast<usize>(static_cast<f64>(total_nodes) / seconds))
+              << std::endl;
 }
